# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Oslava Admin AI Chatbot API
  version: 1.0.0
  description: |
    Production REST API for the standalone Oslava Admin AI Chatbot backend.
    Provides natural-language conversational analytics, operational queries,
    and safe two-phase mutations with explicit admin confirmation.

servers:
  - url: http://localhost:3000
    description: Local Development / Staging Environment
  - url: https://api-chatbot.oslava.com
    description: Production Environment

paths:
  /healthz:
    get:
      summary: Liveness Probe
      description: Returns process uptime and status. No authentication required.
      responses:
        "200":
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "ok" }
                  service: { type: string, example: "oslava-admin-ai" }
                  timestamp: { type: string, format: date-time }
                  uptime: { type: number, example: 124.5 }

  /readyz:
    get:
      summary: Readiness Probe
      description: Checks configuration and persistence readiness without executing expensive AI calls.
      responses:
        "200":
          description: Service is ready to accept traffic
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "ready" }
                  service: { type: string, example: "oslava-admin-ai" }
                  persistence: { type: string, enum: [memory, postgres] }
                  timestamp: { type: string, format: date-time }
        "503":
          description: Service is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /metrics:
    get:
      summary: Operational Metrics
      description: Returns lightweight in-memory counters and latency statistics without PII or secrets.
      responses:
        "200":
          description: Metrics snapshot
          content:
            application/json:
              schema:
                type: object

  /v1/auth/me:
    get:
      summary: Get Authenticated Admin Profile
      description: Returns current caller's profile verified from Supabase JWT and profiles table.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Verified admin profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  role: { type: string, enum: [ADMIN, SUPER_ADMIN] }
                  account_status: { type: string, enum: [ACTIVE] }
                  full_name: { type: string }
                  email: { type: string, nullable: true }
                  phone: { type: string, nullable: true }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/chat/sessions:
    post:
      summary: Create Chat Session
      description: Initializes a new conversational session for the authenticated admin.
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: "#/components/schemas/ChatSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      summary: List User Sessions
      description: Retrieves all active chat sessions owned by the authenticated admin.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChatSession"

  /v1/chat/sessions/{sessionId}:
    get:
      summary: Get Session Details & State
      description: Returns session metadata along with current grounded entity context.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: "#/components/schemas/ChatSession"
                  session_state:
                    $ref: "#/components/schemas/SessionState"
        "404":
          $ref: "#/components/responses/NotFound"

  /v1/chat/sessions/{sessionId}/messages:
    get:
      summary: Get Message History
      description: Returns paginated message history for a session.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChatMessage"
    post:
      summary: Send User Prompt / Turn
      description: Sends a natural language message to the assistant and executes a conversational agent turn.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, minLength: 1 }
      responses:
        "200":
          description: Agent response (standard message, entity selection, or confirmation required)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id: { type: string, format: uuid }
                  response:
                    oneOf:
                      - $ref: "#/components/schemas/StandardMessageResponse"
                      - $ref: "#/components/schemas/EntitySelectionResponse"
                      - $ref: "#/components/schemas/ConfirmationRequiredResponse"
                  session_state:
                    $ref: "#/components/schemas/SessionStateSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/chat/sessions/{sessionId}/action/pending:
    get:
      summary: Get Active Pending Action
      description: Returns the currently unconfirmed action for this session, if any.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Pending action or null
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending_action:
                    nullable: true
                    $ref: "#/components/schemas/PendingActionSummary"

  /v1/chat/actions/{actionId}:
    get:
      summary: Get Action Details
      description: Returns full metadata, arguments, timestamps, and outcome for an action ID.
      security:
        - BearerAuth: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Action record
          content:
            application/json:
              schema:
                type: object
                properties:
                  action:
                    $ref: "#/components/schemas/ActionRecord"

  /v1/chat/actions/{actionId}/confirm:
    post:
      summary: Confirm and Execute Pending Action
      description: Explicitly confirms a pending action. Performs stale-state validation, claims atomically, executes mutation through request-scoped caller JWT, and verifies read-after-write.
      security:
        - BearerAuth: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Action executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  action_id: { type: string, format: uuid }
                  session_id: { type: string, format: uuid }
                  action_type: { type: string }
                  status: { type: string, example: "SUCCEEDED" }
                  display_summary: { type: object }
                  result_summary: { type: object }
                  message: { type: string }
        "400":
          description: Action expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Duplicate confirmation or stale entity state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /v1/chat/actions/{actionId}/cancel:
    post:
      summary: Cancel Pending Action
      description: Cancels a pending action and marks it as CANCELLED without modifying underlying entity data.
      security:
        - BearerAuth: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Action cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  action_id: { type: string, format: uuid }
                  session_id: { type: string, format: uuid }
                  action_type: { type: string }
                  status: { type: string, example: "CANCELLED" }
                  display_summary: { type: object }
                  message: { type: string }

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Access Token provided upon user login.

  responses:
    BadRequest:
      description: Invalid input or validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient permissions (Requires ADMIN or SUPER_ADMIN)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: Operation rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable, request_id]
          properties:
            code:
              type: string
              enum:
                - AUTH_REQUIRED
                - AUTH_INVALID
                - ROLE_FORBIDDEN
                - ACCOUNT_RESTRICTED
                - INVALID_INPUT
                - ENTITY_NOT_FOUND
                - DOMAIN_REJECTED
                - SUPABASE_UNAVAILABLE
                - MODEL_UNAVAILABLE
                - MODEL_RATE_LIMITED
                - MODEL_TIMEOUT
                - MODEL_INVALID_RESPONSE
                - TOOL_LIMIT_EXCEEDED
                - SESSION_NOT_FOUND
                - SESSION_FORBIDDEN
                - ACTION_NOT_FOUND
                - ACTION_FORBIDDEN
                - ACTION_EXPIRED
                - ACTION_ALREADY_RESOLVED
                - ACTION_STALE
                - ACTION_EXECUTION_FAILED
                - ACTION_OUTCOME_UNKNOWN
                - PENDING_ACTION_EXISTS
                - INTERNAL_ERROR
            message: { type: string }
            retryable: { type: boolean }
            request_id: { type: string }

    ChatSession:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        status: { type: string, example: "ACTIVE" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_activity_at: { type: string, format: date-time }

    ChatMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        role: { type: string, enum: [USER, ASSISTANT, SYSTEM_EVENT] }
        content: { type: string }
        created_at: { type: string, format: date-time }

    SessionState:
      type: object
      nullable: true
      properties:
        current_event_id: { type: string, nullable: true }
        current_event_label: { type: string, nullable: true }
        current_worker_id: { type: string, nullable: true }
        current_worker_label: { type: string, nullable: true }
        recent_events: { type: array, items: { type: object } }
        recent_workers: { type: array, items: { type: object } }

    SessionStateSummary:
      type: object
      nullable: true
      properties:
        current_event_id: { type: string, nullable: true }
        current_event_label: { type: string, nullable: true }
        current_worker_id: { type: string, nullable: true }
        current_worker_label: { type: string, nullable: true }

    StandardMessageResponse:
      type: object
      required: [type, content]
      properties:
        type: { type: string, const: "message" }
        content: { type: string }

    EntitySelectionResponse:
      type: object
      required: [type, entityType, question, options]
      properties:
        type: { type: string, const: "entity_selection_required" }
        entityType: { type: string, enum: [event, worker] }
        question: { type: string }
        options:
          type: array
          items:
            type: object
            required: [id, label]
            properties:
              id: { type: string }
              label: { type: string }
              metadata: { type: object }

    ConfirmationRequiredResponse:
      type: object
      required: [type, actionId, actionType, displaySummary, content, expiresAt]
      properties:
        type: { type: string, const: "confirmation_required" }
        actionId: { type: string, format: uuid }
        actionType:
          type: string
          enum:
            - change_worker_category
            - publish_event
            - complete_event
            - close_event
        displaySummary: { type: object }
        content: { type: string }
        expiresAt: { type: string, format: date-time }

    PendingActionSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        action_type: { type: string }
        status: { type: string }
        display_summary: { type: object }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }

    ActionRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        action_type: { type: string }
        status: { type: string }
        arguments: { type: object }
        display_summary: { type: object }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        confirmed_at: { type: string, format: date-time, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        executed_at: { type: string, format: date-time, nullable: true }
        result_summary: { type: object, nullable: true }
        execution_error_code: { type: string, nullable: true }